HashAPI/requirements.txt
flask
requests
blurhash-python